refactored syncing code, moved most of the syncing logic to a new class and unified group syncing
