+++ app/src/main/java/com/mickstarify/zooforzotero/LibraryActivity/Contract.kt
@@ -112,7 +112,7 @@ interface Contract {
         fun deleteAttachment(item: Item)
         fun uploadAttachment(attachment: Item)
         fun getUnfiledItems(): List<Item>
-        fun loadGroup(group: GroupInfo, refresh: Boolean = false)
+        fun startGroupSync(group: GroupInfo, refresh: Boolean = false)
         fun usePersonalLibrary()
         fun getGroupByTitle(groupTitle: String): GroupInfo?
         fun removeFromRecentlyViewed(attachment: Item)
