+++ app/src/main/java/com/mickstarify/zooforzotero/di/component/ApplicationComponent.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import com.mickstarify.zooforzotero.AttachmentManager.AttachmentManagerModel
 import com.mickstarify.zooforzotero.LibraryActivity.LibraryActivityModel
 import com.mickstarify.zooforzotero.SettingsActivity
+import com.mickstarify.zooforzotero.ZoteroAPI.Syncing.SyncManager
 import com.mickstarify.zooforzotero.ZoteroStorage.AttachmentStorageManager
 import com.mickstarify.zooforzotero.ZoteroStorage.ZoteroDB.ZoteroDB
 import com.mickstarify.zooforzotero.di.module.ApplicationModule
@@ -19,5 +20,6 @@ interface ApplicationComponent {
     fun inject(settingsActivity: SettingsActivity)
     fun inject(attachmentManagerModel: AttachmentManagerModel)
     fun inject(attachmentStorageManager: AttachmentStorageManager)
+    fun inject(syncManager: SyncManager)
     fun inject(zoteroDB: ZoteroDB)
 }
